Application ended

+++ src/main/java/com/example/resourcemanagerapp/service/MyUserServiceImpl.java
@@ -1,13 +1,16 @@
 package com.example.resourcemanagerapp.service;
 
 import com.example.resourcemanagerapp.additionalTypes.MyUserType;
+import com.example.resourcemanagerapp.model.MyResource;
 import com.example.resourcemanagerapp.model.MyUser;
+import com.example.resourcemanagerapp.repository.MyResourceRepository;
 import com.example.resourcemanagerapp.repository.MyUserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Optional;
 
 
@@ -15,10 +18,12 @@ import java.util.Optional;
 public class MyUserServiceImpl implements MyUserService{
 
     private final MyUserRepository myUserRepository;
+    private final MyResourceRepository myResourceRepository;
 
     @Autowired
-    public MyUserServiceImpl(MyUserRepository myUserRepository) {
+    public MyUserServiceImpl(MyUserRepository myUserRepository, MyResourceRepository myResourceRepository) {
         this.myUserRepository = myUserRepository;
+        this.myResourceRepository = myResourceRepository;
     }
 
     @Override
@@ -29,8 +34,14 @@ public class MyUserServiceImpl implements MyUserService{
         myUserRepository.save(myUser);
     }
 
+    @Transactional
     @Override
     public void deleteUser(Integer id) {
+        Optional<MyUser> myUser = myUserRepository.findById(id);
+        if(myUser.isEmpty()) return;
+        List<MyResource> userResources = myResourceRepository.findAllByUserId(myUser.get());
+        for(MyResource myResource: userResources)
+            myResourceRepository.deleteById(myResource.getId());
         myUserRepository.deleteById(id);
     }
